+++ record_android/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.0.0
+* chore: **Breaking change** Remove background recording service.
 * chore: Updates minimum supported SDK version to Flutter 3.44/Dart 3.12.
 * chore: Move to AGP 9.x.
 * chore: Move to Kotlin Gradle DSL.